# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organisations/{organisationId}/ai/suggestions": {
      "post": {
        "operationId": "Ai_suggest",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe: a repeated request with the same key and body replays the first response (24h).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "tags": [
          "AI"
        ],
        "x-required-permissions": [
          "ai:use"
        ],
        "summary": "Request an AI suggestion (async). The result is a PROPOSAL awaiting human review.",
        "description": "**Requires permission:** `ai:use`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAiSuggestionRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}/ai/proposals": {
      "get": {
        "operationId": "Ai_list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "description": "Page size"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "`nextCursor` from the previous page"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "SUGGESTED",
                "UNDER_REVIEW",
                "ACCEPTED",
                "REJECTED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "AI"
        ],
        "x-required-permissions": [
          "ai:use"
        ],
        "summary": "List AI proposals",
        "description": "**Requires permission:** `ai:use`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/ai/proposals/{proposalId}": {
      "get": {
        "operationId": "Ai_get",
        "parameters": [
          {
            "name": "proposalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposal"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "AI"
        ],
        "x-required-permissions": [
          "ai:use"
        ],
        "summary": "Get a proposal",
        "description": "**Requires permission:** `ai:use`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/ai/proposals/{proposalId}/review": {
      "post": {
        "operationId": "Ai_review",
        "parameters": [
          {
            "name": "proposalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Under review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposal"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "AI"
        ],
        "x-required-permissions": [
          "ai:approve"
        ],
        "summary": "Take a SUGGESTED proposal into human review (SUGGESTED -> UNDER_REVIEW; requires ai:approve)",
        "description": "**Requires permission:** `ai:approve`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/ai/proposals/{proposalId}/decision": {
      "post": {
        "operationId": "Ai_decide",
        "parameters": [
          {
            "name": "proposalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposal"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "AI"
        ],
        "x-required-permissions": [
          "ai:approve"
        ],
        "summary": "Accept or reject a proposal that is UNDER_REVIEW (requires ai:approve; recorded in workflow history). Accepting applies nothing: the owning module acts through its own authorised service.",
        "description": "**Requires permission:** `ai:approve`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideProposalRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "UK Accounting Platform API",
    "description": "Practice-first accounting, tax and compliance platform. **V0 foundation:** identity, tenancy, RBAC, documents, jobs, audit, outbox events, workflows, tasks, notifications, integration and AI abstractions.\n\nAll tenant data lives under `/organisations/{organisationId}` and is protected by authentication, RBAC, company scope and PostgreSQL row-level security. Errors use RFC 9457 `application/problem+json`.",
    "version": "1.0.0",
    "contact": {},
    "x-api-versioning": "URI versioning (/api/v1). Additive changes only within a version; breaking changes ship as /api/v2."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Opaque session token from `POST /auth/login/bearer`"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uk_session",
        "description": "httpOnly SameSite=Strict cookie set by `POST /auth/login`. Unsafe methods must send an allowed `Origin`."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "urn:uk-platform:error:validation_error"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "instance": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "description": "Echoes `x-request-id`"
          }
        }
      },
...
      "JobAccepted": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "status"
        ],
        "additionalProperties": false
      },
...
      "AiProposal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "payload": {},
          "status": {
            "type": "string",
            "enum": [
              "SUGGESTED",
              "UNDER_REVIEW",
              "ACCEPTED",
              "REJECTED"
            ]
          },
          "provider": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "promptVersion": {
            "type": "string",
            "nullable": true
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "nullable": true,
            "description": "0..1 where the provider supplies one"
          },
          "sourceEvidence": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                }
              },
              "required": [
                "type",
                "id"
              ],
              "additionalProperties": false
            }
          },
          "reviewStartedByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "appliedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "appliedByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "appliedReference": {
            "type": "string",
            "nullable": true
          },
          "workflowInstanceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "requestedByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "decidedByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "decisionComment": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "provider",
          "model",
          "promptVersion",
          "confidence",
          "sourceEvidence",
          "reviewStartedByUserId",
          "appliedAt",
          "appliedByUserId",
          "appliedReference",
          "workflowInstanceId",
          "requestedByUserId",
          "decidedByUserId",
          "decidedAt",
          "decisionComment"
        ],
        "additionalProperties": false
      },
...
      "AiProposalPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "kind": {
                  "type": "string"
                },
                "payload": {},
                "status": {
                  "type": "string",
                  "enum": [
                    "SUGGESTED",
                    "UNDER_REVIEW",
                    "ACCEPTED",
                    "REJECTED"
                  ]
                },
                "provider": {
                  "type": "string",
                  "nullable": true
                },
                "model": {
                  "type": "string",
                  "nullable": true
                },
                "promptVersion": {
                  "type": "string",
                  "nullable": true
                },
                "confidence": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "string"
                    }
                  ],
                  "nullable": true,
                  "description": "0..1 where the provider supplies one"
                },
                "sourceEvidence": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "type",
                      "id"
                    ],
                    "additionalProperties": false
                  }
                },
                "reviewStartedByUserId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "appliedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "appliedByUserId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "appliedReference": {
                  "type": "string",
                  "nullable": true
                },
                "workflowInstanceId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "requestedByUserId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "decidedByUserId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "decidedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "decisionComment": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "kind",
                "status",
                "provider",
                "model",
                "promptVersion",
                "confidence",
                "sourceEvidence",
                "reviewStartedByUserId",
                "appliedAt",
                "appliedByUserId",
                "appliedReference",
                "workflowInstanceId",
                "requestedByUserId",
                "decidedByUserId",
                "decidedAt",
                "decisionComment"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "RequestAiSuggestionRequest": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{2,60}$"
          },
          "input": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "purpose",
          "input"
        ],
        "additionalProperties": false
      },
      "DecideProposalRequest": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "ACCEPT",
              "REJECT"
            ]
          },
          "comment": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "decision"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "400": {
        "description": "Malformed request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "401": {
        "description": "Not authenticated or session expired",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "403": {
        "description": "Authenticated but not permitted (RBAC, segregation of duties, CSRF origin)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "404": {
        "description": "Not found — also returned for resources in other organisations (existence is never revealed)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "409": {
        "description": "Conflict with current state",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "422": {
        "description": "Validation failed (`errors[]` lists fields)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "503": {
        "description": "Dependency unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}